Extract auth helpers
